Make spark use /tmp for local dir.
Spark normally takes the yarn local dir. However, it does this only from the
local dir on the master node. This means all the spark executors on all the
nodes want to talk to the one hod localworkdir used by the master. This commit
changes this to use /tmp

Another solution would have been to set HOD_LOCALWORKDIR in the environments and
use {{HOD_LOCALWORKDIR}} and let spark's templating resolve it later.